refactor(services): découper lms.js par domaine + dédup getClasses (#26)
PR2/3 du chantier dédup #26.

lms.js (466 l., 6 domaines, violation SRP/§5) devient une façade baril de
53 l. composant 6 modules domaine, miroir du split backend LMSDataController :
lmsClasses, lmsMatieres, lmsTeachers, lmsSeances, lmsVisio, lmsNotifications
(tous ≤ 300 l.).

Rétro-compat : lmsService (named + default) reste une façade agrégée → aucun
des 19 appelants n'est modifié. Le code neuf peut importer le service de
domaine directement.

Dédup : lmsService.getClasses (doublon strict de klassciService.getClasses sur
/proxy/classes) retiré ; SeanceManagement repointé vers klassciService.

Tests 159/159 (+ lmsServiceSplit.test.js), contract 28/28, build OK.
